將帳戶餘額與交易金額類型由 double 改為 BigDecimal，提升金融計算精度與安全性

+++ src/main/java/net/javaguides/banking/controller/AccountController.java
@@ -16,6 +16,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -43,9 +44,9 @@ public class AccountController {
     }
 
     @PutMapping("/{id}/deposit")
-    public ResponseEntity<AccountDto> deposit(@PathVariable Long id, @RequestBody Map<String, Double> request) {
+    public ResponseEntity<AccountDto> deposit(@PathVariable Long id, @RequestBody Map<String, BigDecimal> request) {
 
-        Double amount = request.get("amount");
+        BigDecimal amount = request.get("amount");
 
         AccountDto deposit = accountService.deposit(id, amount);
 
@@ -53,8 +54,8 @@ public class AccountController {
     }
 
     @PutMapping("/{id}/withdraw")
-    public ResponseEntity<AccountDto> withdraw(@PathVariable Long id, @RequestBody Map<String, Double> request) {
-        Double amount = request.get("amount");
+    public ResponseEntity<AccountDto> withdraw(@PathVariable Long id, @RequestBody Map<String, BigDecimal> request) {
+        BigDecimal amount = request.get("amount");
         AccountDto accountDto = accountService.withdraw(id, amount);
         return ResponseEntity.status(HttpStatus.OK).body(accountDto);
     }
