+++ src/main/java/net/javaguides/banking/service/AccountService.java
@@ -6,6 +6,7 @@ import net.javaguides.banking.dto.TransferFundDTO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
@@ -14,9 +15,9 @@ public interface AccountService {
 
     AccountDto getAccountById(Long id);
 
-    AccountDto deposit(Long id, Double amount);
+    AccountDto deposit(Long id, BigDecimal amount);
 
-    AccountDto withdraw(Long id, Double amount);
+    AccountDto withdraw(Long id, BigDecimal amount);
 
     Page<AccountDto> getAllAccounts(Pageable pageable);
